Modified most classes, moved functionality out of main class, fixed "factory" class. Can now spawn multiple levels (only collliders a spawn currently need modification of rending) using TAB key for each level. Can "kill" player using ENTER key. Improved (but not fixed) colliders for level objects.

+++ GP Assimp Test/CollisionSystem.h
@@ -124,7 +124,7 @@ public:
 	}
 
 	// create a cube
-	btRigidBody* AddCube(float width, float height, float depth, float x, float y, float z, float mass) {
+	btRigidBody* AddCube(double width, double height, double depth, float x, float y, float z, int mass) {
 		btTransform t;
 		t.setIdentity();
 		t.setOrigin(btVector3(x, y, z));
